feat: Add Microsoft Entra login option with unlock multi-tenancy
The code changes include adding support for Microsoft Entra login by introducing a new icon and updating the login translation files for both English and Japanese languages. Additionally, the necessary environment variables for Microsoft Entra integration have been added to the `.env.example` and `docker-compose.yaml` files.
This change uses OAuth authentication, which also unlocks the multi-tenant functionality that was previously restricted in the Community Edition.

Warning: Using multi-tenant functionality in the Community Edition is a violation of the license.